feat(ObservedUser): Property wrapper to observe a user in cache and maintain it up to date
