--- src/front/macOS/kDriveCore/ServerBridge/Cache/CoherentCacheObservation.swift
@@ -1,28 +1,0 @@
-/*
- Infomaniak kDrive - Desktop
- Copyright (C) 2023-2025 Infomaniak Network SA
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import Combine
-import Foundation
-
-extension AnyPublisher where Output == IndexedUsers, Failure == Never {
-    func userPublisher(for id: Int32) -> AnyPublisher<User?, Never> {
-        map { $0[id] }
-            .removeDuplicates { $0 == $1 }
-            .eraseToAnyPublisher()
-    }
-}
